Update p2000.py
The Capcode's will now ignore the "ignored capcodes", when searching for the region. (I suggest to ignore the 2 general groupcodes that are send with most of the messages (002029568, 002029569))
If these 2 capcodes are ignored, the region check is 100% accurate (if the capcode is known in the database).

I tried to configure a friendly name to the sensor, but ended up, hardcoding it :-( There is possible a better way.